Move gates and flags out of jujud
The following moves all the gate and housing related code ouf of
the cmd/jujud package.
