+++ agent/engine/valueworker_test.go
@@ -1,4 +1,4 @@
-// Copyright 2015 Canonical Ltd.
+// Copyright 2023 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
 package engine_test
@@ -9,7 +9,7 @@ import (
 	"github.com/juju/worker/v3"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/cmd/jujud/agent/engine"
+	"github.com/juju/juju/agent/engine"
 )
 
 type ValueWorkerSuite struct {
