+++ cmd/jujud/agent/caasoperator/manifolds.go
@@ -17,12 +17,12 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	coreagent "github.com/juju/juju/agent"
+	"github.com/juju/juju/agent/engine"
 	"github.com/juju/juju/api"
 	caasoperatorapi "github.com/juju/juju/api/agent/caasoperator"
 	"github.com/juju/juju/api/base"
 	"github.com/juju/juju/api/client/charms"
 	"github.com/juju/juju/caas/kubernetes/provider/exec"
-	"github.com/juju/juju/cmd/jujud/agent/engine"
 	"github.com/juju/juju/core/machinelock"
 	"github.com/juju/juju/core/status"
 	"github.com/juju/juju/juju/sockets"
