+++ cmd/jujud/agent/machine/stateflag.go
@@ -7,7 +7,7 @@ import (
 	"github.com/juju/worker/v3"
 	"github.com/juju/worker/v3/dependency"
 
-	"github.com/juju/juju/cmd/jujud/agent/engine"
+	"github.com/juju/juju/agent/engine"
 )
 
 // isControllerFlagManifold returns a dependency.Manifold that indicates
