+++ worker/gate/flag_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/juju/worker/v3/workertest"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/cmd/jujud/agent/engine"
+	"github.com/juju/juju/agent/engine"
 	"github.com/juju/juju/worker/gate"
 )
