+++ worker/metrics/spool/manifold.go
@@ -13,7 +13,7 @@ import (
 	"gopkg.in/tomb.v2"
 
 	"github.com/juju/juju/agent"
-	"github.com/juju/juju/cmd/jujud/agent/engine"
+	"github.com/juju/juju/agent/engine"
 )
 
 // MetricRecorder records metrics to a spool directory.
