+++ worker/retrystrategy/manifold.go
@@ -10,8 +10,8 @@ import (
 	"github.com/juju/worker/v3/dependency"
 
 	"github.com/juju/juju/agent"
+	"github.com/juju/juju/agent/engine"
 	"github.com/juju/juju/api/base"
-	"github.com/juju/juju/cmd/jujud/agent/engine"
 	"github.com/juju/juju/rpc/params"
 )
