+++ worker/toolsversionchecker/manifold.go
@@ -12,10 +12,10 @@ import (
 	"github.com/juju/worker/v3/dependency"
 
 	"github.com/juju/juju/agent"
+	"github.com/juju/juju/agent/engine"
 	apiagent "github.com/juju/juju/api/agent/agent"
 	"github.com/juju/juju/api/base"
 	"github.com/juju/juju/api/controller/agenttools"
-	"github.com/juju/juju/cmd/jujud/agent/engine"
 )
 
 // ManifoldConfig defines the names of the manifolds on which a Manifold will depend.
